Move remote command handlers to picard/remotecommands
- Tagger class is big enough
- most methods are using Tagger object which is available through QtCore.QCoreApplication.instance()
- ParseItemsToLoad class is only used by remote command handlers (tests modified accordingly)
- TestMessageParsing -> TestParseItemsToLoad (as it does that)
